feature(homepage): create homepage
Create homepage where user will navigate to before choosing to log in or sign up.
Changes include addition of the UI directory for all UI files, creation of index.html
file (and associated style sheets), a css (for style sheets) , a js (for scripts), and
an images (for images) folder inside the UI directory.