refactor(files): refactor app files to real world behavior
This app is my Andela Developer Challenge for the Blended Learning
Program. It was required we submit on an autograder app which was
required that I make changes to my app. It is not dynamic as a human
and so complying is the only way forward, not explaining my reasons
for decisions.
This commit undo many of the alterations before it in order for the
app to behave like a real world app would, make restrictions to resources
where necessary depending on the permission level.
It also fixes tests failing due to the changes made to specifications

[Finishes #167359050, Delivers #167359050, Fixes #167317231)]